refactor: 整理 LLM 重试策略工具模块
将共享退避工具统一归入 utils 模块，并补全重试决策与等待时间解析的文档说明。
