+++ test/kernel/llm/test_policy_utils.py
@@ -3,7 +3,7 @@ import math
 import pytest
 
 from src.kernel.llm import LLMAPIError, LLMRateLimitError, LLMTimeoutError
-from src.kernel.llm.policy import backoff
+from src.kernel.llm.policy import utils
 
 
 @pytest.mark.parametrize(
@@ -13,9 +13,9 @@ from src.kernel.llm.policy import backoff
 def test_retry_delay_uses_exponential_full_jitter(
     monkeypatch, ordinal, expected_upper_bound
 ):
-    monkeypatch.setattr(backoff.random, "uniform", lambda low, high: high / 2)
+    monkeypatch.setattr(utils.random, "uniform", lambda low, high: high / 2)
 
-    delay = backoff.retry_delay(
+    delay = utils.retry_delay(
         model={"retry_interval": 2},
         error=TimeoutError("timeout"),
         retry_ordinal=ordinal,
@@ -35,9 +35,9 @@ def test_retry_delay_uses_exponential_full_jitter(
     ],
 )
 def test_retryable_errors_use_model_retry_interval(monkeypatch, error):
-    monkeypatch.setattr(backoff.random, "uniform", lambda low, high: high)
+    monkeypatch.setattr(utils.random, "uniform", lambda low, high: high)
 
-    delay = backoff.retry_delay(
+    delay = utils.retry_delay(
         model={"retry_interval": 7.5},
         error=error,
         retry_ordinal=2,
@@ -50,9 +50,9 @@ def test_retry_after_takes_priority_without_jitter(monkeypatch):
     def fail_uniform(low, high):
         raise AssertionError("Retry-After 不应使用 jitter")
 
-    monkeypatch.setattr(backoff.random, "uniform", fail_uniform)
+    monkeypatch.setattr(utils.random, "uniform", fail_uniform)
 
-    delay = backoff.retry_delay(
+    delay = utils.retry_delay(
         model={"retry_interval": 2},
         error=LLMRateLimitError("limited", retry_after=12.5),
         retry_ordinal=3,
@@ -62,26 +62,26 @@ def test_retry_after_takes_priority_without_jitter(monkeypatch):
 
 
 def test_retry_delay_caps_retry_after_and_backoff(monkeypatch):
-    monkeypatch.setattr(backoff.random, "uniform", lambda low, high: high)
+    monkeypatch.setattr(utils.random, "uniform", lambda low, high: high)
 
-    retry_after_delay = backoff.retry_delay(
+    retry_after_delay = utils.retry_delay(
         model={"retry_interval": 1},
         error=LLMRateLimitError("limited", retry_after=120),
         retry_ordinal=1,
     )
-    backoff_delay = backoff.retry_delay(
+    backoff_delay = utils.retry_delay(
         model={"retry_interval": 30},
         error=TimeoutError("timeout"),
         retry_ordinal=4,
     )
 
-    assert retry_after_delay == backoff.MAX_RETRY_DELAY_SECONDS
-    assert backoff_delay == backoff.MAX_RETRY_DELAY_SECONDS
+    assert retry_after_delay == utils.MAX_RETRY_DELAY_SECONDS
+    assert backoff_delay == utils.MAX_RETRY_DELAY_SECONDS
 
 
 @pytest.mark.parametrize("value", [-1, "invalid", float("nan"), float("inf")])
 def test_retry_delay_handles_invalid_base_interval(value):
-    delay = backoff.retry_delay(
+    delay = utils.retry_delay(
         model={"retry_interval": value},
         error=TimeoutError("timeout"),
         retry_ordinal=1,
